A Study On coupon Based Promotional Activities
The analysis revealed that coupon promotion which has a direct relationship with income level, Trustworthy 
brands. The Correlation coefficient shows that coupon promotion and income level and 
trustworthy brands are positively related. The regression analysis denotes a significant positive 
impact of coupon promotion and reward points on income level and trustworthy brands. After 
gathering and interpreting data, conclusions were drawn on the basis of observations gathered 
during the analysis.